move redirects to separate file
