+++ docs/_docset.yml
@@ -9,26 +9,6 @@ subs:
   serverless-short: Serverless
   ece:   "Elastic Cloud Enterprise"
   eck:   "Elastic Cloud on Kubernetes"
-redirects:
-  'testing/redirects/4th-page.md': 'testing/redirects/5th-page.md'
-  'testing/redirects/6th-page.md':
-  'testing/redirects/first-page-old.md':
-    to: 'testing/redirects/second-page.md'
-    anchors:
-      'old-anchor': 'active-anchor'
-      'removed-anchor':
-  'testing/redirects/second-page-old.md':
-    many:
-      - to: "testing/redirects/second-page.md"
-        anchors:
-          "aa": "zz"
-          "removed-anchor":
-      - to: "testing/redirects/third-page.md"
-        anchors:
-          "bb": "yy"
-  'testing/redirects/third-page.md':
-    anchors:
-      'removed-anchor':
 toc:
   - file: index.md
   - hidden: developer-notes.md
