+++ docs/_redirects.yml
@@ -1,0 +1,20 @@
+redirects:
+  'testing/redirects/4th-page.md': 'testing/redirects/5th-page.md'
+  'testing/redirects/6th-page.md':
+  'testing/redirects/first-page-old.md':
+    to: 'testing/redirects/second-page.md'
+    anchors:
+      'old-anchor': 'active-anchor'
+      'removed-anchor':
+  'testing/redirects/second-page-old.md':
+    many:
+      - to: "testing/redirects/second-page.md"
+        anchors:
+          "aa": "zz"
+          "removed-anchor":
+      - to: "testing/redirects/third-page.md"
+        anchors:
+          "bb": "yy"
+  'testing/redirects/third-page.md':
+    anchors:
+      'removed-anchor':
